requirements.txt
Flask
Flask_Login
Flask_PyMongo
protobuf
Requests
Werkzeug
google-generativeai